Users can register and login(required to access website)
In the file forms.py, a new class for buttons is created for user registration.

In the file register.html, the webpage template for registration is created.

In the file models.py, the User class is modified (new method) so the class can
set passwords and check them. Including to this, the User class extends UserMixin
so it can have access to is_authenticated(), is_active(), is_anonymous(), and
get_id(). Finaly, a new method to load user is created.

In the file routes.py, the login() method is modified so it takes advantage
of flask_login methods. A new method for logout() is created to log out users.
These methods are created so a login is required when accessing my website (
and also to take advatage the decorator @login_required)

Finaly, the file register.html is created for the registration web page.