Users have a profile page and avatar from gravatar.com
The file routes.py was modified. A new route was create (/user/<username>) that
created the template for each user profile. Including to this a new function
with the decorator @app.before_request was created to update the last_seen
variable (a User variable) anytime a specified user profile is requested. Finaly,
a new route was created (/edit_profile) to help users edit their own personal
profile.

The file user.html was created for rendering the user profile.

The file base.html was modified to have the profile internal link.

The file _post.html was created to create the html code for a users post.

The file edit_profile.html was created to render the profile editing page.

The file models.py was modified. A new function (avatar(self, size)) was created
for the User class. This function creates the avatar image. Including to this
the Use class has two new global variables (about_me and last_seen), which will
contain an about me info and the last date the user was seen, respectively.

The file forms.py was modified. A new class (EditProfileForm) was created to
contain the needed input fields for editing a users profile.